refactor: move RatingSelectRadio to own component
