+++ src/App/RatingSelectCard/index.tsx
@@ -1,56 +1,25 @@
 import React from "react";
-import { Rating } from "../@model";
-import styles from "./styles.module.scss";
+import { AppModel, Rating, updateRatingOrNoop } from "../@model";
+import RatingSelectRadio from "./RatingSelectRadio";
 
-export default function RatingSelectCard(props: {
-  selectedRating: Rating | null;
-  onSelect: (selectedRating: Rating) => void;
-  name: string;
+export function RatingSelectCardView(props: {
+  model: AppModel;
+  updateModel: (updateFn: (oldModel: AppModel) => AppModel) => void;
 }) {
-  const { selectedRating, onSelect, name } = props;
-  const allRatings = [1, 2, 3, 4, 5] as Rating[];
-  return (
-    <div className={styles.mainContainer}>
-      {allRatings.map((i) => {
-        return (
-          <RatingRadioButton
-            key={i}
-            isSelected={i === selectedRating}
-            onSelect={onSelect}
-            name={name}
-            rating={i}
-          />
-        );
-      })}
-    </div>
-  );
-}
+  const { model, updateModel } = props;
+  const handleSelect = (selectedRating: Rating) => {
+    const updateFn = (oldModel: AppModel) =>
+      updateRatingOrNoop(oldModel, selectedRating);
+    updateModel(updateFn);
+  };
 
-function RatingRadioButton(props: {
-  isSelected: boolean;
-  onSelect: (rating: Rating) => void;
-  name: string;
-  rating: Rating;
-}) {
-  const { isSelected, onSelect, name, rating } = props;
   return (
-    <label
-      htmlFor={`${name}-${rating}`}
-      className={`${styles.radioLabel} ${
-        isSelected && styles.selectedRadioLabel
-      }`}
-    >
-      <input
-        type="radio"
-        name={name}
-        id={`${name}-${rating}`}
-        className={styles.radioInput}
-        checked={isSelected}
-        onChange={() => {
-          onSelect(rating);
-        }}
-      />
-      {rating}
-    </label>
+    <form>
+      <RatingSelectRadio
+        selectedRating={model.rating}
+        onSelect={handleSelect}
+        name={"rating"}
+      ></RatingSelectRadio>
+    </form>
   );
 }
